refactor: move mainWindow.ts into its own dir
